requirements.txt
# Core NLP libraries
# nltk
# keybert
# rake-nltk
# gensim

# Machine learning
# scikit-learn
# pandas
# numpy

# OpenAI integration
openai

# Utility packages
python-dotenv
# transformers
# sentence-transformers

# Web utilities (optional, for potential future enhancements)
# beautifulsoup4
# requests